Initial system for #10
The particle system is in its severe infancy! But we have it. Bloom will be coming soon, as well as actually documenting the code.

Currently we have a FireParticleSystem and an ImpactParticleSystem. A few others will be coming, they will have defaults set to match what they are, but they will be customizable. If you want to use a FireParticleSystem for some magical spawning effect it will have the flexibility to be used as such.

I think im going to shoot for 10-12 specific systems all dynamic enough to cover most of our bases.