Prep work for #3
Started a world file, this will include the rooms and doors that connect them. Rooms are broken into 50x50 pixel chunks.

Player is temporarily also 50x50 for testing, he will soon be facing the mouse instead of facing the direction of keyboard movement.

All particle systems have been temporarily removed from the main.py as they all need to be reworked to reflect the new perspective and subsequent new gravity!

Lots of work to be done!