Rough first pass on player rotation
Reading into some math I ended up prototyping more of the particle system and a bullet system. Its a absolute clusterfuck of code so its not written in here, fixing player rotation to the center of the object instead of the top-left corner comes next. Then fancy projectiles and spells.