#6 Added menu options to menu.py
Added five options to the top left section of the menu that will in the future, change what is displayed in the other two sections when the text is clicked. Also need to clean up my code a bit. Now that I've gotten a better grasp on what I'm trying to do, I am going to condense it using dictionaries.